interfaces HorizontalBlindService, PriceCatalog, UserPage created
