+++ src/Main.java
@@ -1,11 +1,12 @@
 import view.UserPage;
+import view.UserPageImpl;
 
 public class Main {
     public static void main(String[] args) {
 
-        UserPage userPage = new UserPage();
+        UserPage userPageImpl = new UserPageImpl();
 
-        userPage.showBlindCost();
+        userPageImpl.showBlindCost();
 
     }
 }
