+++ src/service/PriceCatalog.java
@@ -1,20 +1,9 @@
 package service;
 
-import java.util.HashMap;
 import java.util.Map;
 
-public class PriceCatalog {
+public interface PriceCatalog {
 
-    private final Map<Integer, Double> colorMap = new HashMap<>();
-
-    public Map<Integer, Double> initDate() {
-        colorMap.put(201, 8.8);
-        colorMap.put(202, 10.4);
-
-        return colorMap;
-    }
-
-    public double getColorPrice(int color) {
-        return colorMap.get(color);
-    }
+    Map<Integer, Double> initDate();
+    double getColorPrice(int color);
 }
