+++ src/testing/TestHorizontalBlindService.java
@@ -2,8 +2,8 @@ package testing;
 
 import org.junit.jupiter.api.Test;
 import model.Blind;
-import service.HorizontalBlindService;
-import service.PriceCatalog;
+import service.HorizontalBlindServiceImpl;
+import service.PriceCatalogImpl;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -12,12 +12,12 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class TestHorizontalBlindService {
 
-    HorizontalBlindService horizontalBlindService = new HorizontalBlindService();
+    HorizontalBlindServiceImpl horizontalBlindServiceImpl = new HorizontalBlindServiceImpl();
 
     @Test
     void testCalculateCostOfBlinds() {
         Blind blind = new Blind();
-        PriceCatalog catalog = new PriceCatalog();
+        PriceCatalogImpl catalog = new PriceCatalogImpl();
         catalog.initDate();
         blind.setBlindsWidth(2000);
         blind.setBlindsHeight(4000);
@@ -25,7 +25,7 @@ public class TestHorizontalBlindService {
         Map<Integer, Double> map = new HashMap<>();
         map.put(201, 8.8);
         map.put(202, 10.4);
-        assertEquals(1056, horizontalBlindService.calculateCostOfBlinds(blind, catalog));
+        assertEquals(1056, horizontalBlindServiceImpl.calculateCostOfBlinds(blind, catalog));
 
     }
 }
